perf(db): reader/writer split, profile_gen in-place updates, ServerStat i64
Database concurrency: from_pool() opens two turso connections — writer
(Mutex) for writes, reader (RwLock) for reads — enabling concurrent
reads without writer contention. Removed sync conn() helper entirely.

Profile mutation: replace full reload_profiles() after every add/
update/delete with profile_gen counter + upsert_profile_row() in-place
upsert, cutting redundant DB round trips.

ServerStat: today_up/today_down/total_up/total_down widened from i32
to i64 to prevent overflow on large traffic accumulations.

Subscription upsert: changed prepare() to prepare_cached() for the
three inner INSERT/UPDATE statements, reusing prepared statement handles.

Ping cancel: narrowed cancel_ping_batch to only abort status=queued,
preventing killing in-flight fast pings.

Borrow fix: clone prev_id and drop filtered borrow before mutating
self in move_profile handler.

+++ crates/xray-tui-db/src/convert.rs
@@ -74,10 +74,10 @@ impl ServerStat {
     pub(crate) fn from_row(row: &turso::Row) -> turso::Result<Self> {
         Ok(Self {
             profile_id: row.get::<String>(ServerStatCol::ProfileId as usize)?,
-            today_up: row.get::<Option<i32>>(ServerStatCol::TodayUp as usize)?,
-            today_down: row.get::<Option<i32>>(ServerStatCol::TodayDown as usize)?,
-            total_up: row.get::<Option<i32>>(ServerStatCol::TotalUp as usize)?,
-            total_down: row.get::<Option<i32>>(ServerStatCol::TotalDown as usize)?,
+            today_up: row.get::<Option<i64>>(ServerStatCol::TodayUp as usize)?,
+            today_down: row.get::<Option<i64>>(ServerStatCol::TodayDown as usize)?,
+            total_up: row.get::<Option<i64>>(ServerStatCol::TotalUp as usize)?,
+            total_down: row.get::<Option<i64>>(ServerStatCol::TotalDown as usize)?,
             last_updated: row.get::<Option<String>>(ServerStatCol::LastUpdated as usize)?,
         })
     }
